add route handler to add an event to a group by its id

+++ backend/db/migrations/20230119001035-create-event.js
@@ -40,7 +40,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       price: {
-        type: Sequelize.INTEGER
+        type: Sequelize.REAL
       },
       startDate: {
         type: Sequelize.DATE
